encoding working
