+++ .gitignore
@@ -74,4 +74,4 @@ _testmain.go
 *.test
 *.prof
 
-
+*.out
