+++ encode.go
@@ -5,11 +5,8 @@ import (
 	"fmt"
 	"reflect"
 	"strconv"
-	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
-	"golang.org/x/exp/constraints"
 )
 
 // Marshaler is the interface implemented by objects that can marshal themselves into
@@ -403,341 +400,6 @@ func marshal(v interface{}, flags encodeFlags) (*dynamodb.AttributeValue, error)
 // 	}
 // }
 
-func encodeSliceSet(rt /* []T */ reflect.Type, flags encodeFlags) (encodeFunc, error) {
-	switch {
-	case rt.Elem().Implements(rtypeTextMarshaler):
-		return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-			ss := make([]*string, 0, rv.Len())
-			for i := 0; i < rv.Len(); i++ {
-				tm := rv.Index(i).Interface().(encoding.TextMarshaler)
-				text, err := tm.MarshalText()
-				if err != nil {
-					return nil, err
-				}
-				if flags&flagOmitEmptyElem != 0 && len(text) == 0 {
-					continue
-				}
-				ss = append(ss, aws.String(string(text)))
-			}
-			if len(ss) == 0 {
-				return nil, nil
-			}
-			return &dynamodb.AttributeValue{SS: ss}, nil
-		}, nil
-	}
-	switch rt.Elem().Kind() {
-	// NS
-	case reflect.Int, reflect.Int64, reflect.Int32, reflect.Int16, reflect.Int8:
-		return encodeSliceNS[int64]((reflect.Value).Int, strconv.FormatInt), nil
-	case reflect.Uint, reflect.Uint64, reflect.Uint32, reflect.Uint16, reflect.Uint8:
-		return encodeSliceNS[uint64]((reflect.Value).Uint, strconv.FormatUint), nil
-	case reflect.Float64, reflect.Float32:
-		return encodeSliceNS[float64]((reflect.Value).Float, formatFloat), nil
-
-	// SS
-	case reflect.String:
-		return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-			ss := make([]*string, 0, rv.Len())
-			for i := 0; i < rv.Len(); i++ {
-				s := rv.Index(i).String()
-				if flags&flagOmitEmptyElem != 0 && s == "" {
-					continue
-				}
-				ss = append(ss, aws.String(s))
-			}
-			if len(ss) == 0 {
-				return nil, nil
-			}
-			return &dynamodb.AttributeValue{SS: ss}, nil
-		}, nil
-
-	// BS
-	case reflect.Slice:
-		if rt.Elem().Elem().Kind() == reflect.Uint8 {
-			return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-				bs := make([][]byte, 0, rv.Len())
-				for i := 0; i < rv.Len(); i++ {
-					b := rv.Index(i).Bytes()
-					if flags&flagOmitEmptyElem != 0 && len(b) == 0 {
-						continue
-					}
-					bs = append(bs, b)
-				}
-				if len(bs) == 0 {
-					return nil, nil
-				}
-				return &dynamodb.AttributeValue{BS: bs}, nil
-			}, nil
-		}
-	}
-
-	return nil, fmt.Errorf("dynamo: invalid type for set: %v", rt)
-}
-
-func encodeMapM(rt reflect.Type, flags encodeFlags) (encodeFunc, error) {
-	var keyString func(k reflect.Value) (string, error)
-	if ktype := rt.Key(); ktype.Implements(rtypeTextMarshaler) {
-		keyString = func(k reflect.Value) (string, error) {
-			tm := k.Interface().(encoding.TextMarshaler)
-			txt, err := tm.MarshalText()
-			if err != nil {
-				return "", fmt.Errorf("dynamo: marshal map: key error: %v", err)
-			}
-			return string(txt), nil
-		}
-	} else if ktype.Kind() == reflect.String {
-		keyString = func(k reflect.Value) (string, error) {
-			return k.String(), nil
-		}
-	} else {
-		return nil, fmt.Errorf("dynamo marshal: map key must be string, have %v", rt)
-	}
-
-	subflags := flagNone
-	if flags&flagAllowEmptyElem != 0 {
-		subflags |= flagAllowEmpty | flagNull
-		// child containers of a map also have the allowEmptyElem behavior
-		// i.e. lists inside a map or maps inside a map
-		subflags |= flagAllowEmptyElem
-	} else if flags&flagOmitEmptyElem != 0 {
-		subflags |= flagOmitEmpty
-	}
-
-	valueEnc, err := encoderFor(rt.Elem(), subflags)
-	if err != nil {
-		return nil, err
-	}
-
-	return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-		if rv.IsNil() {
-			if flags&flagAllowEmpty != 0 {
-				return &dynamodb.AttributeValue{M: map[string]*dynamodb.AttributeValue{}}, nil
-			}
-			if flags&flagNull != 0 {
-				return &dynamodb.AttributeValue{NULL: aws.Bool(true)}, nil
-			}
-			return nil, nil
-		}
-
-		avs := make(map[string]*dynamodb.AttributeValue, rv.Len())
-
-		iter := rv.MapRange()
-		for iter.Next() {
-			v, err := valueEnc(iter.Value(), subflags)
-			if err != nil {
-				return nil, err
-			}
-			if v == nil {
-				return nil, nil
-			}
-
-			kstr, err := keyString(iter.Key())
-			if err != nil {
-				return nil, err
-			}
-
-			avs[kstr] = v
-		}
-
-		if flags&flagOmitEmpty != 0 && len(avs) == 0 {
-			return nil, nil
-		}
-
-		return &dynamodb.AttributeValue{M: avs}, nil
-	}, nil
-}
-
-func encodeMapSet(rt /* map[T]bool | map[T]struct{} */ reflect.Type, flags encodeFlags) (encodeFunc, error) {
-	truthy := truthy(rt)
-	useBool := truthy.Kind() == reflect.Bool
-	if !truthy.IsValid() {
-		return nil, fmt.Errorf("dynamo: cannot marshal type %v into a set (value type of map must be ~bool or ~struct{})", rt)
-	}
-
-	if rt.Key().Implements(rtypeTextMarshaler) {
-		return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-			length := rv.Len()
-			ss := make([]*string, 0, length)
-			// strs := make([]string, length)
-			iter := rv.MapRange()
-			for iter.Next() {
-				if useBool && !iter.Value().Equal(truthy) {
-					continue
-				}
-				text, err := iter.Key().Interface().(encoding.TextMarshaler).MarshalText()
-				if err != nil {
-					return nil, err
-				}
-				if flags&flagOmitEmptyElem != 0 && len(text) == 0 {
-					continue
-				}
-				str := string(text)
-				ss = append(ss, &str)
-			}
-			return &dynamodb.AttributeValue{SS: ss}, nil
-		}, nil
-	}
-
-	switch rt.Key().Kind() {
-	// NS
-	case reflect.Int, reflect.Int64, reflect.Int32, reflect.Int16, reflect.Int8:
-		return encodeMapNS[int64](truthy, (reflect.Value).Int, strconv.FormatInt), nil
-	case reflect.Uint, reflect.Uint64, reflect.Uint32, reflect.Uint16, reflect.Uint8:
-		return encodeMapNS[uint64](truthy, (reflect.Value).Uint, strconv.FormatUint), nil
-	case reflect.Float32, reflect.Float64:
-		return encodeMapNS[float64](truthy, (reflect.Value).Float, formatFloat), nil
-
-	// SS
-	case reflect.String:
-		return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-			ss := make([]*string, 0, rv.Len())
-			iter := rv.MapRange()
-			for iter.Next() {
-				if useBool && !iter.Value().Equal(truthy) {
-					continue
-				}
-				s := iter.Key().String()
-				if flags&flagOmitEmptyElem != 0 && s == "" {
-					continue
-				}
-				ss = append(ss, aws.String(s))
-			}
-			if len(ss) == 0 {
-				return nil, nil
-			}
-			return &dynamodb.AttributeValue{SS: ss}, nil
-		}, nil
-
-	// BS
-	case reflect.Array:
-		if rt.Key().Elem().Kind() == reflect.Uint8 {
-			size := rt.Key().Len()
-			return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-				bs := make([][]byte, 0, rv.Len())
-				key := make([]byte, size)
-				keyv := reflect.ValueOf(key)
-				iter := rv.MapRange()
-				for iter.Next() {
-					if useBool && !iter.Value().Equal(truthy) {
-						continue
-					}
-					reflect.Copy(keyv, iter.Key())
-					bs = append(bs, key)
-				}
-				if len(bs) == 0 {
-					return nil, nil
-				}
-				return &dynamodb.AttributeValue{BS: bs}, nil
-			}, nil
-		}
-	}
-
-	return nil, fmt.Errorf("dynamo: invalid type for set: %v", rt)
-}
-
-func encodeN[T constraints.Integer | constraints.Float](get func(reflect.Value) T, format func(T, int) string) encodeFunc {
-	return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-		str := format(get(rv), 10)
-		return &dynamodb.AttributeValue{N: &str}, nil
-	}
-}
-
-func encodeSliceNS[T constraints.Integer | constraints.Float](get func(reflect.Value) T, format func(T, int) string) encodeFunc {
-	return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-		ns := make([]*string, 0, rv.Len())
-		for i := 0; i < rv.Len(); i++ {
-			n := get(rv.Index(i))
-			if flags&flagOmitEmptyElem != 0 && n == 0 {
-				continue
-			}
-			str := format(n, 10)
-			ns = append(ns, &str)
-		}
-		if len(ns) == 0 {
-			return nil, nil
-		}
-		return &dynamodb.AttributeValue{NS: ns}, nil
-	}
-}
-
-func encodeMapNS[T constraints.Integer | constraints.Float](truthy reflect.Value, get func(reflect.Value) T, format func(T, int) string) encodeFunc {
-	useBool := truthy.Kind() == reflect.Bool
-	return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-		ns := make([]*string, 0, rv.Len())
-		iter := rv.MapRange()
-		for iter.Next() {
-			if useBool && !iter.Value().Equal(truthy) {
-				continue
-			}
-			n := get(iter.Key())
-			if flags&flagOmitEmptyElem != 0 && n == 0 {
-				continue
-			}
-			str := format(n, 10)
-			ns = append(ns, &str)
-		}
-		if len(ns) == 0 {
-			return nil, nil
-		}
-		return &dynamodb.AttributeValue{NS: ns}, nil
-	}
-}
-
-func encodeSet(rt /* []T | map[T]bool | map[T]struct{} */ reflect.Type, flags encodeFlags) (encodeFunc, error) {
-	switch rt.Kind() {
-	case reflect.Slice:
-		return encodeSliceSet(rt, flags)
-	case reflect.Map:
-		return encodeMapSet(rt, flags)
-	}
-
-	return nil, fmt.Errorf("dynamo: marshal: invalid type for set %s", rt.String())
-}
-
-func encodeList(rt reflect.Type, flags encodeFlags) (encodeFunc, error) {
-	// lists CAN be empty
-	subflags := flagNone
-	if flags&flagOmitEmptyElem == 0 {
-		// unless "omitemptyelem" flag is set, include empty/null values
-		// this will preserve the position of items in the list
-		subflags |= flagAllowEmpty | flagNull
-	}
-	if flags&flagAllowEmptyElem != 0 {
-		// child containers of a list also have the allowEmptyElem behavior
-		// e.g. maps inside a list
-		subflags |= flagAllowEmptyElem
-	}
-
-	valueEnc, err := encoderFor(rt.Elem(), subflags)
-	if err != nil {
-		return nil, err
-	}
-
-	return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-		avs := make([]*dynamodb.AttributeValue, 0, rv.Len())
-		for i := 0; i < rv.Len(); i++ {
-			innerVal := rv.Index(i)
-			av, err := valueEnc(innerVal, flags|subflags)
-			if err != nil {
-				return nil, err
-			}
-			if av == nil {
-				if flags&flagOmitEmptyElem != 0 {
-					continue
-				}
-				av = nullAV
-			}
-			if av != nil {
-				avs = append(avs, av)
-			}
-		}
-		if flags&flagOmitEmpty != 0 && len(avs) == 0 {
-			return nil, nil
-		}
-		return &dynamodb.AttributeValue{L: avs}, nil
-	}, nil
-}
-
 var emptyStructType = reflect.TypeOf(struct{}{})
 
 func marshalSliceNoOmit(values []interface{}) ([]*dynamodb.AttributeValue, error) {
@@ -829,46 +491,6 @@ func fieldInfo(field reflect.StructField) (name string, flags encodeFlags) {
 // 	str := strconv.FormatInt(unix, 10)
 // 	return &dynamodb.AttributeValue{N: &str}, nil
 // }
-
-func encodeUnixTime(rt reflect.Type) encodeFunc {
-	switch rt {
-	case rtypeTimePtr:
-		return encode2[*time.Time](func(t *time.Time, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-			if t == nil || t.IsZero() {
-				return nil, nil
-			}
-			str := strconv.FormatInt(t.Unix(), 10)
-			return &dynamodb.AttributeValue{N: &str}, nil
-		})
-	case rtypeTime:
-		return encode2[time.Time](func(t time.Time, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-			if t.IsZero() {
-				return nil, nil
-			}
-			str := strconv.FormatInt(t.Unix(), 10)
-			return &dynamodb.AttributeValue{N: &str}, nil
-		})
-	}
-	panic(fmt.Errorf("not time type: %v", rt))
-}
-
-func encode2[T any](fn func(T, encodeFlags) (*dynamodb.AttributeValue, error)) func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-	return func(rv reflect.Value, flags encodeFlags) (*dynamodb.AttributeValue, error) {
-		if !rv.IsValid() || !rv.CanInterface() {
-			return nil, nil
-		}
-
-		if (rv.Kind() == reflect.Pointer || rv.Kind() == reflect.Interface) && rv.IsNil() {
-			if flags&flagNull != 0 {
-				return nullAV, nil
-			}
-			return nil, nil
-		}
-
-		v := rv.Interface().(T)
-		return fn(v, flags)
-	}
-}
 
 type isZeroer interface {
 	IsZero() bool
